Refactor: Move CSV validation to utils module.
- I created a new `utils` folder and `utils/utils.py` file.
- I moved the `validate_csv` function and its pandas import from `app.py` to `utils/utils.py`.
- I updated `app.py` to import and use `validate_csv` from the new location.
- I updated `README.md` to reflect these structural changes.

This change improves code organization by separating utility functions from the main application logic.
