+++ programacao/GaleriaAnimes/README.md
@@ -1,5 +1,6 @@
 ## Features
-- Includes a CSV validation function in `app.py` to check if `data/dados.csv` is empty using the pandas library.
+- Includes a CSV validation function in `utils/utils.py` to check if `data/dados.csv` is empty using the pandas library.
+- A `utils` module (`utils/utils.py`) has been added to house common utility functions.
 
 ## Dependencies
 - pandas
